### 2025-04-02 #7977 @editorialbot
[PRE REVIEW]: HiMAP:  Hidden Markov for Advanced Prognostics
**Submitting author:** @mariana-sc (<a href="http://orcid.org/0009-0008-0315-8686">Mariana Salinas-Camus</a>)
**Repository:** https://github.com/GroupiSP/himap
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
