+++ paper.bib
@@ -35,6 +35,7 @@
 
 @article{guo2019review,
   title={A review on prognostics methods for engineering systems},
+  doi={10.1109/TR.2019.2957965},
   author={Guo, J. and Li, Z. and Li, M.},
   journal={IEEE Transactions on Reliability},
   volume={69},
@@ -46,6 +47,7 @@
 
 @article{da2020remaining,
   title={Remaining useful lifetime prediction via deep domain adaptation},
+  doi={10.1016/j.ress.2019.106682},
   author={da Costa, Paulo Roberto de Oliveira and Ak{\c{c}}ay, Alp and Zhang, Yingqian and Kaymak, Uzay},
   journal={Reliability Engineering \& System Safety},
   volume={195},
@@ -56,6 +58,7 @@
 
 @article{li2020data,
   title={Data alignments in machinery remaining useful life prediction using deep adversarial neural networks},
+  doi={10.1016/j.knosys.2020.105843},
   author={Li, Xiang and Zhang, Wei and Ma, Hui and Luo, Zhong and Li, Xu},
   journal={Knowledge-Based Systems},
   volume={197},
@@ -66,6 +69,7 @@
 
 @inproceedings{vollert2021challenges,
   title={Challenges of machine learning-based RUL prognosis: A review on NASA's C-MAPSS data set},
+  doi={10.1109/ETFA45728.2021.9613682},
   author={Vollert, S. and Theissler, A.},
   booktitle={2021 26th IEEE International Conference on Emerging Technologies and Factory Automation (ETFA)},
   pages={1--8},
