+++ docs/quick_start.rst
@@ -3,6 +3,12 @@ Quick Start Guide: Prognostics using C-MAPSS
 
 This guide provides a step-by-step process for loading the C-MAPSS dataset, creating a Hidden Markov Model (HMM) and a Hidden Semi-Markov Model (HSMM), and performing prognostics using your model.
 
+.. warning::
+    This example will NOT run with the PyPi install, since the package does not 
+    contain the example data. To get all of the required files, please build
+    the entire GitHub repository, following the respective instructions.
+        
+
 .. contents:: Table of Contents
    :depth: 3
    :local:
@@ -10,8 +16,8 @@ This guide provides a step-by-step process for loading the C-MAPSS dataset, crea
 Using the classic Hidden Markov Model (HMM) for prognostics
 ===========================================================
 
-Step 1: Load the C-MAPSS
-------------------------
+Step 1: Load the C-MAPSS data example
+-------------------------------------
 
 The data used corresponds to sensor 11 data from the FD001 sub-dataset of C-MAPSS, discretized into 20 values for easier use with HMMs. The training dataset consists of 80 run-to-failure degradation histories, while the testing dataset contains 20 run-to-failure degradation histories.
 
@@ -28,7 +34,7 @@ To load the data, use the following Python code:
    obs_state_len = 5  # Number of times the failure value is repeated in the data
 
    # Load training and testing data
-   seqs_train, seqs_test = load_cmapss(f_value=f_value, obs_state_len=obs_state_len)
+   seqs_train, seqs_test = load_data_cmapss(f_value=f_value, obs_state_len=obs_state_len)
 
 The ``f_value`` refers to the failure value (21), and ``obs_state_len`` indicates how many times this failure value is repeated in the training data. This ensures that the model learns a consistent failure pattern. To better understand the degradation process, visualize a few degradation histories from the training set. Below, you can see 8 example trajectories, showing how sensor readings change over time. Since this dataset comes from one failure mode (FD001), the degradation trends in training and testing are similar.  
 
@@ -44,8 +50,7 @@ Next, define the Hidden Markov Model (HMM). The number of observation symbols is
 
 .. code-block:: python
 
-   n_states = 5
-   hmm_c = HMM(n_states=n_states, n_obs_symbols=f_value)
+   hmm_c = HMM(n_obs_symbols=f_value)
 
 You can also provide ``n_iter`` as an input, which specifies the maximum number of iterations for training the model.
 
@@ -56,32 +61,33 @@ To determine the optimal number of states, you can use the Bayesian Information 
 
 .. code-block:: python
 
-   hmm_c, bic = hmm_c.fit_bic(seqs_train, states=list(np.arange(2, n_states + 2)))
+   hmm_c, bic = hmm_c.fit_bic(seqs_train, states=list(np.arange(2, 15)))
 
-This tests models with 2 to 7 states (you can adjust the range) and returns the best one based on BIC.
+This tests models with 2 to 15 states (you can adjust the range) and returns the best one based on BIC.
 
-If you already know the number of states, train your model directly:
+If you already know the number of hidden states that you want to use, you can define them during the instantiation of the model with the ``n_states`` parameter and train your model directly:
 
 .. code-block:: python
-
+   
+   hmm_c = HMM(n_states=5, n_obs_symbols=f_value)
    hmm_c.fit(seqs_train)
 
 Step 4: Perform Prognostics
 ---------------------------
 
-Once trained, use the HMM to predict the Remaining Useful Life (RUL):
[…]
