### 2025-08-27 #7977 @mariana-sc
Hello @pdebuyl, thank you for your comments.

1. We are going to change now the license that is OSI approved.
2. In terms of the Python requirement, we are doing tests to check until which version the software is stable. We would update the requirements in the README by the end of the week.
3. There is no procedure for automated tests.

Please let me know if anything else needs to be clarified, and if the changes that we are going to make are sufficient.
