### 2025-08-27 ca6dd6b
Create CITATION.cff
