### 2025-08-27 5aa6cc9
Update pyproject.toml for licensing
