+++ pyproject.toml
@@ -31,8 +31,9 @@ keywords = ["Hidden Markov Models", "Prognostics"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Programming Language :: Python"
+    "License :: OSI Approved :: Apache Software License",
 ]
-license = { text = "CC-BY-4.0" }
+license = { text = "Apache-2.0" }
 
 [project.urls]
 Documentation = "https://himap.readthedocs.io/en/latest/"
