### 2025-08-27 15e8f63
licensing update
