### 2025-08-27 32ebcec
licensing update_minor fix
