+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "himap"
-version = "1.1.0"
+version = "1.2.0"
 requires-python = ">=3.9, <3.13"
 dependencies = [
     "matplotlib>=3.7.1",
@@ -31,6 +31,11 @@ keywords = ["Hidden Markov Models", "Prognostics"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Programming Language :: Python",
+    "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
 license = "Apache-2.0"
 license-files = ["LICENSE"]
