### 2025-09-08 #7977 @mariana-sc
Hello @pdebuyl, hope you are doing well. Is it all in good shape to start the review?
