### 2025-11-22 #9491 @AnnikaStein
-> the README has more / different installation processes than the readthedocs page

-> a bit confusing, does the software only work if (and only if) the special dataset is shipped with with the software? Can this not be resolved by separating code and data?
-> idea for improvement: check python version requirements, is there a reason this only works with `3.9.16`?
-> as mentioned above, would like to have same requirements / installation guidelines in README and readthedocs, or at least mention that the full guidelines are stated in the other place (readthedocs only mentions pip install, nothing else, while README seems more complete and different options)
-> one example is given in api docs, however, unclear what steps users need to follow if they want to use another dataset; do they need to write another function to read in different data? Do the input data files need to follow a specific format / structure?
-> no automatic tests, functionality can only be tested for FD001 sub-dataset of C-MAPSS
-> guidelines for 1./2./3. missing in README
-> at the read the docs page, replace "dummy citation" with what is mentioned in the main repo

-> add a comparison to other packages used in the field; are there other implementations of the HMM / HSMM models available, and how does HiMAP compare in terms of functionality, flexibility, performance?
-> minor editing comment: abbreviations for HMMs, HSMMs are introduced multiple times within same section, use abbreviations after first mention
-> add citations for examples in lines 25f. and 27f.
-> add citation for Viterbi-decoded state sequences (if I understood correctly, the new features use these sequences  implicitly and do not introduce them for the first time, so need a citation)
