### 2025-12-12 GroupiSP/himap#8 @robmoss
Include example data in published package
Hi @GroupiSP , I've started my [JOSS review](https://github.com/openjournals/joss-reviews/issues/9491) and saw that in the Quick Start Guide it states:

> This example will NOT run with the PyPi install, since the package does not contain the example data. To get all of the required files, please build the entire GitHub repository, following the respective instructions.

I had to remind myself how this works, but the good news is that it's reasonably simple to include the example data in the package so that examples will work when installing `himap` from PyPI.

If you add the following lines to `pyproject.toml`, the CSV files will be included in the source and wheel distributions:

```toml
[tool.setuptools]
include-package-data = true
packages.find.namespaces = true
package-data."himap.example_data" = ["*.csv"]
```

These data files can then be loaded in `himap/utils.py` without relying on hard-coded paths or modifying `sys.path`, by using [`pkgutil.get_data()`](https://docs.python.org/3/library/pkgutil.html#pkgutil.get_data):

```py
import io
import pkgutil

# ...

train_data_str = pkgutil.get_data('himap', 'example_data/train_FD001_disc_20_mod.csv')
train_data_file = io.BytesIO(train_data_str)
train = pd.read_csv(train_data_file, sep=';')

test_data_str = pkgutil.get_data('himap', 'example_data/test_FD001_disc_20_mod.csv')
test_data_file = io.BytesIO(test_data_str)
test = pd.read_csv(test_data_file, sep=';')
```

This allows the `main.py` script to function as intended from any working directory, as long as the `himap` package is installed.
