### 2025-12-12 GroupiSP/himap#10 @robmoss
Additional instructions/guidelines
As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/9491), I see that there are test cases in `himap/test_base.by` but there are no instructions for how to run them. Could you please add brief instructions to the README and/or the online documentation?

Please also consider adding guidelines for third parties wishing to contribute to the software, report issues or problems with the software, and seek support. This might be as simple as adding a few sentences to the README, or adding a separate CONTRIBUTING.md file (here is [one such example](https://github.com/desResLab/LINFA/blob/master/CONTRIBUTING.md)).
