### 2025-12-12 GroupiSP/himap#11 @robmoss
Brief description of C-MAPSS data sets
As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/9491) I've been working through the examples in `main.py` and the [Quick Start Guide](https://himap.readthedocs.io/en/latest/quick_start.html), which is really nice and clear — congratulations!

I have a two minor suggestions for this guide:

1. Could you please briefly describe the C-MAPSS data, for people (such as me) who are not aware of it?
2. Could you please add the necessary `import` statement(s) for the code examples (to import `load_data_cmapss`, `HMM`, `GaussianHSMM`)
