### 2025-12-12 GroupiSP/himap#12 @robmoss
JOSS paper comments
As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/9491), I've read through the article. It's clear, well written, and easy to understand; thank you for writing a nice draft. I have a few suggestions about extra information/details that I think would be useful:

- Are there papers that could be referenced to support each of the following two sentences in the first paragraph of the **Statement of need**:

    > For example, in aerospace, advanced prognostics assist in predicting component failures in aircraft, thereby reducing in-flight risks and preventing costly delays.

    > In the energy sector, these methods enable the continuous monitoring of turbine and battery health, optimizing efficiency and extending operational lifespans.

- In the **Statement of need**, could you please add a little more detail about what this package provides? Is the main contribution (a) the `HMM` and `HSMM` base classes; (b) the implementation of tailored fitting and sampling methods for these kinds of models; (c) the novel prognostic measure; (d) all of the above; or (e) something else I've missed?

- I couldn't see any description of how this packages compares to other existing packages. Does this package address a novel gap for which there are no existing packages? Or are there existing packages that are broadly similar in scope or domain? Perhaps the authors could briefly compare this package to, e.g., `scikit-learn` and/or `PyMC`?

- Please consider including the two example RUL plots from the Quick Start Guide in the paper itself — with a common y-axis scale — to give the reader an idea of what the outputs look like and how using a more advanced model can result in better predictions. It wouldn't require much additional text (maybe only a single extra sentence) and I think it would follow on nicely from the current final paragraph.
