### 2025-12-12 GroupiSP/himap#13 @robmoss
Understanding the output files
As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/9491), I've run the example script:

```sh
python3 main.py --hsmm True
```

It took over 3 hours on my laptop, and generated the following outputs:

- Plots of 20 trajectories in `results/figures`, which don't necessarily appear to get better (should they be iterative improvements?)
- Four dictionaries in JSON format that I'm pretty sure contain the lower/mean/upper bounds for each of the 20 trajectories, and a much larger file that defines the full PDF at each time-step for each trajectory
- A CSV file `(`results/df_results.csv`) that contains several summary statistics for each of the 20 trajectories, and for the average across all 20 trajectories (should this be interpreted as an ensemble forecast?)
- What appears to be the model state, stored in binary format in `results/models/hsmm.txt`, presumably so that the model can be applied to other data sets?

Can the authors please let me know where I have correctly understood the contents and purpose of each output?
