### 2025-12-12 #9491 @pdebuyl
Hello @robmoss thanks a lot for the review and the detailed suggestions for the package.

@mariana-sc you can start addressing the issues raised by @robmoss already.
