### 2026-01-06 GroupiSP/himap#13 @mariana-sc
Hi @robmoss, thanks a lot for your feedback. We will start now with implementing the changes.
About your questions here:

- There shouldn't be iterative improvements since the 20 test trajectories that you see plotted are different, but they have similar behavior because the engines went through the same profile of operation. The iteration to improve the results is done in the training process but the plots are only done once the training process is finished.
- Correct
- It shouldn't be interpreted as an ensemble forecast since each trajectory is independent from each other. However, we included the average across all 20 trajectories to have an insight on how the model performs across engines that went under the same operational profile.
- Correct

We will include these explanations in docs, to improve the explanation of the package.
