### 2026-01-07 986fd29
update paper with more references and update quick start guide
