+++ paper.md
@@ -30,13 +30,17 @@ Prognostics, the science of predicting systems' future health, performance, and 
 
 # Statement of need
 
-Modern systems in critical industries, such as aerospace and energy, are often used under different operational conditions, with limited or no labeled data available for training. These systems frequently lack comprehensive physical models to accurately describe degradation processes, making it challenging to predict future failures [@guo2019review]. Therefore, advanced prognostics, which are defined here as providing reliable RUL predictions under such conditions, are essential for optimizing maintenance schedules, reducing downtime, and improving system reliability. For example, in aerospace, advanced prognostics assist in predicting component failures in aircraft, thereby reducing in-flight risks and preventing costly delays. In the energy sector, these methods enable the continuous monitoring of turbine and battery health, optimizing efficiency and extending operational lifespans.
+Modern systems in critical industries, such as aerospace and energy, are often used under different operational conditions, with limited or no labeled data available for training. These systems frequently lack comprehensive physical models to accurately describe degradation processes, making it challenging to predict future failures [@guo2019review]. Therefore, advanced prognostics, which are defined here as providing reliable RUL predictions under such conditions, are essential for optimizing maintenance schedules, reducing downtime, and improving system reliability. For example, in aerospace, advanced prognostics assist in predicting component failures in aircraft, thereby reducing in-flight risks and preventing costly delays [@prognostics2023aircraft]. In the energy sector, these methods enable the continuous monitoring of wind turbines [@LEITE20181917] and battery health [@hu2020battery], optimizing efficiency and extending operational lifespans.
 
 While state-of-the-art Deep Learning (DL) models, such as Long Short-Term Memory (LSTM) networks and Convolutional Neural Networks (CNNs), have shown results with high accuracy, they require large labeled datasets, are sensitive to environmental uncertainties, and struggle to generalize when operational conditions deviate from training scenarios [@vollert2021challenges], [@da2020remaining], [@li2020data].
 
 Stochastic models like Hidden Markov Models (HMMs) [@rabiner] and Hidden Semi-Markov Models (HSMMs) [@yu] offer a robust alternative for advanced prognostics. By treating RUL as a random variable, these models inherently address the uncertainties in degradation processes and adapt to changes in operational conditions. Their probabilistic foundation makes them particularly suited for real-world applications where labeled failure data is sparse or unavailable. Moreover, their ability to model the stochastic nature of degradation processes ensures reliable predictions, even under varying and unpredictable conditions.
 
-HiMAP is a repository that implements Hidden Markov Models (HMMs) and Hidden Semi-Markov Models (HSMMs) specifically designed for prognostics applications. Each model is implemented as a dedicated Python class, designed for seamless integration into diverse workflows. These classes provide essential methods such as `decode`, for inferring the most likely sequence of hidden states; `fit`, for parameter learning; and `sample`, for generating synthetic sequences. Beyond these functionalities, the repository introduces advanced features for calculating RUL using a novel prognostic measure [@phme_conference]. By leveraging Viterbi-decoded state sequences, this measure produces a probability density function (pdf) of RUL, enabling reliable predictions.
[…]
