+++ docs/quick_start.rst
@@ -19,12 +19,25 @@ Using the classic Hidden Markov Model (HMM) for prognostics
 Step 1: Load the C-MAPSS data example
 -------------------------------------
 
-The data used corresponds to sensor 11 data from the FD001 sub-dataset of C-MAPSS, discretized into 20 values for easier use with HMMs. The training dataset consists of 80 run-to-failure degradation histories, while the testing dataset contains 20 run-to-failure degradation histories.
+The C-MAPSS (Commercial Modular Aero-Propulsion System Simulation) dataset is a widely used benchmark in Prognostics and Health Management (PHM), providing simulated degradation data for aircraft turbofan engines operating under various conditions. It contains multivariate time-series measurements from multiple sensors collected over the entire operational life of each engine, from healthy operation to system failure. 
+
+The FD001 sub-dataset represents a single operating condition with a single fault mode, making it well suited for developing and evaluating data-driven approaches to degradation modeling and remaining useful life estimation. The C-MAPSS dataset also includes additional sub-datasets: FD002, with six operating conditions and one fault mode; FD003, with one operating condition and two fault modes; and FD004, with six operating conditions and two fault modes. 
+
+For the purposes of the example case study, only data from sensor 11 of the FD001 sub-dataset is considered. This sensor data is discretized into 20 values to facilitate its use with Hidden Markov Models (HMMs). Each run-to-failure degradation history corresponds to the time-series data collected from a single engine, starting from healthy operation and ending at system failure. The training dataset consists of 80 such degradation histories, while the testing dataset contains 20 degradation histories
 
 The dataset includes two files:
 
 - ``train_FD001_disc_20_mod.csv`` for training
 - ``test_FD001_disc_20_mod.csv`` for testing
+
+
+First, ensure you have the necessary imports:
+
+.. code-block:: python
+
+   import himap
+   from utils import *
+   from himap.base import HMM, GaussianHSMM
 
 To load the data, use the following Python code:
 
@@ -81,7 +94,7 @@ Once the model is trained, you can use the defined prognostic module to predict 
 
    hmm_c.prognostics(seqs_test, plot_rul=True, get_metrics=True)
 
-This function will generate and save RUL plots in a ``figures`` folder and also save a CSV file containing the following performance metrics:
+This function will generate and save RUL plots in a ``figures`` folder, each figure corresponds to the RUL predictions of each test tracjectory. Also, the function saves a CSV file containing the following performance metrics:
 
 - **RMSE**: Measures prediction accuracy (lower is better)
 - **Coverage**: Indicates how well the true RUL values fall within the confidence intervals (ideal = 1)
@@ -89,7 +102,7 @@ This function will generate and save RUL plots in a ``figures`` folder and also 
 
 Additionally, RUL probability distributions (PDFs) for each time step are saved in the ``dictionary`` folder, along with confidence intervals.
 
-Below is an example of the RUL prediction results:
+Below there is an example of the RUL prediction results. The x-axis corresponds to the operational time and the y-axis to the RUL (which has the same units as the operational time). The true RUL is defined as the time remaining before failure, computed using the known failure time of the system and decreasing linearly as the system operates. The predicted RUL represents the model’s estimate of this remaining lifetime based on the observed sensor data. The confidence intervals indicate the uncertainty associated with the prediction, providing a range of plausible RUL values around the estimate, with wider intervals reflecting higher uncertainty.
 
 .. image:: _images/hmm_RUL_plot_traj_19.png
    :align: center
