+++ himap/main.py
@@ -4,7 +4,6 @@
 import sys
 import os
 
-sys.path.append(os.path.join(os.path.dirname(__file__)))
 
 import warnings
 
@@ -12,8 +11,8 @@ warnings.filterwarnings(action="ignore")
 warnings.simplefilter(action="ignore", category=FutureWarning)
 warnings.simplefilter(action="ignore", category=UserWarning)
 
-from utils import *
-from base import GaussianHSMM, HMM
+from .utils import *
+from .base import GaussianHSMM, HMM
 import argparse
 
 
