### 2026-01-26 GroupiSP/himap#14 @thanoskont
Testing
robmoss:
  As part of my https://github.com/openjournals/joss-reviews/issues/9491, I see that there are test cases in himap/test_base.by but there are no instructions for how to run them. Could you please add brief instructions to the README and/or the online documentation?
