### 2026-01-27 #9491 @AnnikaStein
> Hi @AnnikaStein, thanks for the feedback. Could you please clarify to which examples do you refer with this "add citations for examples in lines 25f. and 27f."?. I don't know if its my editor but line 25 and 27 are located before the paper begins.

Hi @mariana-sc I was reading the paper draft itself, which comes with the line numbers on the left. It's about the real world use cases
