### 2026-01-27 #9491 @AnnikaStein
actually, I think I now understand where the confusion about the line numbers comes from

in paper.md indeed there are much fewer lines

I was referring to the article proof pdf (so the rendered paper file after conversion from markdown)

the sentences in question are:

> For example, in aerospace, advanced prognostics assist in predicting component failures in aircraft, thereby reducing in-flight risks and preventing costly delays. In the energy sector, these methods enable the continuous monitoring of turbine and battery health, optimizing efficiency and extending operational lifespans.
