### 2026-01-27 97ddde1
add contributing.md, update reference in main docs, include load other datasets in quick start docs
