+++ paper.md
@@ -34,11 +34,11 @@ Modern systems in critical industries, such as aerospace and energy, are often u
 
 While state-of-the-art Deep Learning (DL) models, such as Long Short-Term Memory (LSTM) networks and Convolutional Neural Networks (CNNs), have shown results with high accuracy, they require large labeled datasets, are sensitive to environmental uncertainties, and struggle to generalize when operational conditions deviate from training scenarios [@vollert2021challenges], [@da2020remaining], [@li2020data].
 
-Stochastic models like Hidden Markov Models (HMMs) [@rabiner] and Hidden Semi-Markov Models (HSMMs) [@yu] offer a robust alternative for advanced prognostics. By treating RUL as a random variable, these models inherently address the uncertainties in degradation processes and adapt to changes in operational conditions. Their probabilistic foundation makes them particularly suited for real-world applications where labeled failure data is sparse or unavailable. Moreover, their ability to model the stochastic nature of degradation processes ensures reliable predictions, even under varying and unpredictable conditions.
+Stochastic models like HMMs [@rabiner] and HSMMs [@yu] offer a robust alternative for advanced prognostics. By treating RUL as a random variable, these models inherently address the uncertainties in degradation processes and adapt to changes in operational conditions. Their probabilistic foundation makes them particularly suited for real-world applications where labeled failure data is sparse or unavailable. Moreover, their ability to model the stochastic nature of degradation processes ensures reliable predictions, even under varying and unpredictable conditions.
 
-HiMAP is a repository that implements Hidden Markov Models (HMMs) and Hidden Semi-Markov Models (HSMMs) specifically designed for prognostics applications. To the best of the authors’ knowledge, it is the first package to provide such models explicitly tailored to prognostics rather than to generic time-series analysis. For example, the packages `hmmlearn` and `hmms` implement continuous and discrete HMMs that can be used for a wide variety of applications, but they are not targeted to a specific discipline. As a result, they do not include the assumptions typically required for prognostics, and it is left to the user to determine how to use the estimated hidden states to compute the RUL. In a similar way, the package `PyMC` offers a wide and well-documented variety of stochastic and Bayesian models that can be applied to time-series analysis and survival analysis (a discipline closely related to prognostics), but none of these models are specifically designed for the prognostic task.
+HiMAP is a repository that implements HMMs and HSMMs specifically designed for prognostics applications. To the best of the authors’ knowledge, it is the first package to provide such models explicitly tailored to prognostics rather than to generic time-series analysis. For example, the packages `hmmlearn` and `hmms` implement continuous and discrete HMMs that can be used for a wide variety of applications, but they are not targeted to a specific discipline. As a result, they do not include the assumptions typically required for prognostics, and it is left to the user to determine how to use the estimated hidden states to compute the RUL. In a similar way, the package `PyMC` offers a wide and well-documented variety of stochastic and Bayesian models that can be applied to time-series analysis and survival analysis (a discipline closely related to prognostics), but none of these models are specifically designed for the prognostic task.
 
[…]
