+++ CONTRIBUTING.md
@@ -1,0 +1,52 @@
+# Contributing
+
+Thank you for your interest in contributing.
+
+This project currently accepts **bug fixes and maintenance changes only**.
+
+## Reporting Bugs
+
+If you find a bug, please open an issue and include:
+
+- A clear description of the problem  
+- Steps to reproduce it  
+- Expected and actual behaviour  
+- Relevant logs, screenshots, or error messages  
+- Details about your environment, if relevant  
+
+Please check existing issues before opening a new one.
+
+---
+
+## Pull Requests
+
+Pull requests are welcome for bug fixes and maintenance improvements.
+
+Guidelines:
+
+1. Fork the repository  
+2. Create a branch for your change  
+3. Make your fix  
+4. Commit with clear, descriptive messages  
+5. Open a pull request against the `main` branch  
+
+Please ensure that:
+
+- Changes are focused and minimal  
+- Existing tests pass  
+- Documentation is updated if behaviour changes  
+- Tests are added or updated where applicable  
+
+---
+
+## Commit Messages
+
+- Use the present tense  
+- Keep messages concise and descriptive  
+- Reference related issues where relevant  
+
+---
+
+## Questions
+
+If something is unclear, open an issue for discussion.
