### 2026-01-28 e4979ef
Implemented tests, changed import logic, included example_data to the distribution
