+++ requirements.txt
@@ -1,8 +1,0 @@
-numpy>=1.23.5
-scipy>=1.11.0 
-scikit_learn==1.3.0 
-matplotlib>=3.7.1
-pandas==1.5.3 
-tqdm==4.65.0 
-numba==0.60 
-Cython==0.29.35
