### 2026-01-28 647d345
updated CONTRIBUTING.md for running pytests
