+++ requirements.txt
@@ -1,0 +1,8 @@
+matplotlib>=3.7.1,
+numpy>=1.23.5,
+scikit_learn>=1.3.0,
+scipy>=1.10.0,
+tqdm>=4.65.0,
+numba>=0.60,
+pandas>=1.5.3,
+Cython>=0.29.35
