### 2026-01-28 33f7b0a
Revise installation steps and usage examples
Updated installation instructions and example usage in README.
