+++ README.md
@@ -45,25 +45,25 @@ pip install himap
 ```
 
 >[!Note]
->To install the package, you need `Python=3.9.16`
+>To install the package, you need `Python>=3.9`
 
 
 ### Option 2: Install from Source Code
 
 If you prefer to install HiMAP directly from the source, follow these steps:
 
-1. Create an Anaconda environment and activate it.
+1. Create a virtual environment and activate it. (This example will be demonstrated with Anaconda, but it is not required.)
 
   Step 1a
 
 ```
-conda create -n himap python=3.9.16
+conda create -n himap_env python=3.12
 ```
 
   Step 1b
 
 ```
-conda activate himap
+conda activate himap_env
 ```
 
 
@@ -102,12 +102,14 @@ python setup_cython.py build_ext --inplace
 
 
 ```
-../hsmm_dev/
+../root/
       └── LICENSE
       └── README.md
       └── requirements.txt
+      └── ...
+
     
-      ├── hmm/                                                          -- Required
+      ├── himap/                                                        -- Required
           └── ab.py                                                     -- Required
           └── base.py                                                   -- Required
           └── main.py                                                   -- Required
@@ -115,7 +117,8 @@ python setup_cython.py build_ext --inplace
           └── smoothed.pyd                                              -- Required
           └── utils.py                                                  -- Required
           
-          ├── cython_build/                                             -- Required      
+          ├── cython_build/                                             -- Required
+              └── __init__.py                                           -- Required      
               └── fwd_bwd.pyx                                           -- Required
               └── setup.py                                              -- Required
 
@@ -132,6 +135,8 @@ python setup_cython.py build_ext --inplace
 
 ## Example
 
+
+### Run from the rebuild repo
 
 To describe how to train and use the HMM and HSMM models, we show an example below. To run the code from the Anaconda terminal with default values, go to the `himap` directory and run the `main.py` file via the commands:
 
@@ -159,7 +164,16 @@ python main.py --mc_sampling True
 
 See the `main.py` file for different existing variables and options.
 
-### Results
+### Run from the installed package
+
+The example can be also run via the distribution. After installing the package in the virtual environment you can run
+
+```
+python -m himap.main
+```
+and use the same arguments as previously for the different example options.
+
+## Results
 
 The results are saved inside the directory `../himap/results/`
 
