### 2026-01-29 6edccc6
Update image paths in paper.md for RUL plots
