### 2026-01-29 7f01b70
deleted obsolete main.py and images
