+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "himap"
-version = "1.2.1.dev6"
+version = "1.3.0"
 requires-python = ">=3.9, <3.13"
 dependencies = [
     "matplotlib>=3.7.1",
