### 2026-01-29 GroupiSP/himap#14 @thanoskont
@robmoss, Thank you for your suggestion to explain the testing procedure. We implemented further integration and smoke testing, and included a CONTRIBUTING file where the testing procedure is fully explained.
