### 2026-01-29 GroupiSP/himap#10 @thanoskont
@robmoss As per your request, we implemented a CONTRIBUTING.md file, explaining the ways to contribute as well as how to run the newly implemented testing.

Thank you.
