### 2026-01-29 GroupiSP/himap#8 @thanoskont
@robmoss, including the example data in the release is indeed a very clever and easy way to avoid rebuilding the repo just so the example can be run.
Thank you for the suggestion.
