### 2026-01-29 cdd68d1
Add citation section to README
Added citation information and DOIs for the HiMAP package.
