### 2026-01-29 GroupiSP/himap#12 @mariana-sc
Hi Rob, thanks for the comment on the paper.

- We added references for the examples in the statement of need.
- In the statement of need, in the description of the HiMAP package we have detailed what the package provides, such as the HMM and HSMM, the methods, and the prognostic measure.
- We also added a comparison with other packages such as hmmlearn, hmms, and PyMC, and explained what HiMAP offers that is different from them.
- Thank you for the suggestion. We included the graphs and indeed it gives the reader a better idea of what they are going to get by using the package.
