### 2026-01-29 GroupiSP/himap#11 @mariana-sc
We updated the Quick Start Guide. Now it has a more complete description of the dataset and the code examples include the import statements.
