### 2026-01-29 #9491 @mariana-sc
Hi @AnnikaStein , we have addressed all of your review feedback and implemented the requested changes.
In particular, the package now supports Python versions 3.9 and higher. In the Quick Start Guide, we added instructions for using the package with different datasets. We also included a CONTRIBUTING file that explains the testing procedure.

Hi @robmoss, we have responded to the issues you opened, and all changes are now visible in the main branch. There was some overlap between your comments and Annika’s, particularly regarding testing and paper references, which we addressed together.

We hope these changes sufficiently address your feedback.
