### 2026-02-01 #9491 @robmoss
@mariana-sc @pdebuyl I've completed my review, thanks very much to the authors for incorporating my feedback.

I have one additional suggestion: the axis and legend labels in the manuscript figures are extremely small — can you please make them larger so that they are a similar size to the main text?

Apart from that, the authors have addressed all of my comments 👍
