### 2026-02-03 #9491 @AnnikaStein
@mariana-sc Thanks for the updated paper and repo!

with this all my items are addressed, but I echo what @robmoss said regarding size of the figure labels in comparison to the main text.
