### 2026-02-03 ffbe997
Update RUL plot images in paper.md
