### 2026-02-03 #9491 @mariana-sc
Thanks both @robmoss and @AnnikaStein for your quick answers. We have just changed the font size labels of the RUL plots both in the paper and in the docs.
@pdebuyl please let us know what are the next steps.
