### 2026-02-03 #9491 @pdebuyl
Thanks @robmoss @AnnikaStein for the reviews and @mariana-sc for the updates! I will read the updated version of the manuscript and proceed to the next editorial steps in the coming days. The procedure changed a bit at the turn of the year, so I can't point to the old version, but the current one is in this page: https://joss.readthedocs.io/en/latest/editing.html
