### 2026-04-01 #9491 @pdebuyl
Hello @mariana-sc

disclaimer: the text below is what stood cached in my reply box for which I didn't hit the "Comment" button. Sorry, it will probably feel a bit disjoint.

"text that stood in the cached web page"

I ran the software and found a few minor issues with it:

1. The git repository contains `__pycache__` directories. Those are cache directories and should not be registered in git. Can you remove them? You can also add `__pycache__` to the `.gitignore` file to avoid re-committing such files in the future.
2. `himap` stores models in files whose extension is `.txt`, but the files are pickle files. This can cause confusion and another extension would be more suited.

In the paper, a section name on line 42, such as "Description of the models" or "Overview of `himap`'s features" would be welcome. This part of the article is not about the statement of need anymore.

"newer text"

Some issue that I found hard to articulate as I am testing the code after the reviewers: it is not obvious how the CLI tool is only usable for the built-in model or not. The HMM / HSMM models are used to generate samples, but if the CLI only handles a default set of transition rates it should be more explicit in the documentation and article.
