### 2026-04-02 #9491 @mariana-sc
Hi @pdebuyl,
I think there may be a small misunderstanding about the role of the CLI. The CLI is intended as a quick way to run the provided examples, rather than as a general tool within the package itself. Users who want to run the model on their own datasets would do so by using the package directly in their own code (following the info from the docs). Is this what you were referring to?
The remaining minor issues have already been addressed. Once we are aligned on the CLI aspect, we’ll commit all the changes together.
