### 2026-04-02 a892df0
Added pycache and build files to .gitignore, fixed model saving format ambiguity, exposed path parameters to customize saving path, added a return option to HSMM.prongositcs and HMM.prognostics methods
