+++ himap/base.py
@@ -4,6 +4,7 @@
 import sys
 import os
 import warnings
+
 warnings.simplefilter(action='ignore', category=FutureWarning)
 warnings.simplefilter(action='ignore', category=DeprecationWarning)
 warnings.simplefilter(action='ignore', category=RuntimeWarning)
@@ -24,6 +25,7 @@ from himap.plot import *
 
 try:
     from himap.cython_build import fwd_bwd as core
+
     _CORE_IMPORT_ERROR = None
 except Exception as e:
     core = None
@@ -38,7 +40,7 @@ class HSMM:
     """
 
     def __init__(self, n_states=2, n_durations=5, n_iter=20, tol=1e-2, left_to_right=False, obs_state_len=None,
-                 f_value=None, random_state=None, name=""):
+                 f_value=None, random_state=None, name="", results_parent_path=None):
 
         """
         Parameters
@@ -61,6 +63,9 @@ class HSMM:
             Seed for reproducibility.
         name : str, optional
             Name of the model. Defaults to "hsmm" if not provided.
+        results_parent_path : str, optional
+            The path to create the himap_results directory tree where the results (models, figures, dictionaries,
+            performance metrics) are saved
         """
 
         if not n_states >= 2:
@@ -81,7 +86,12 @@ class HSMM:
             self.last_observed = False
 
         # create the folders
-        create_folders()
+        create_folders(results_parent_path)
+
+        if results_parent_path is None:
+            self.results_path = os.path.join(os.getcwd(), "himap_results")
+        else:
+            self.results_path = os.path.join(results_parent_path, "himap_results")
 
         self.max_len = None
         self.n_states = n_states
@@ -698,10 +708,10 @@ class HSMM:
 
             # save the previous version of the model prior to updating
             if save_iters:
-                path = os.path.join(os.getcwd(), 'results', 'models', f'{self.name}_iter_{str(itera + 1)}.txt')
-                with open(path, 'wb') as f:
-                    pickle.dump(self, f)
-                    print(f"\nModel saved at {path}")
+                model_name = self.name
+                self.name = f'{self.name}_iter_{str(itera + 1)}'
+                self.save_model()
+                self.name = model_name
 
             # emission parameters re-estimation
             weight = mean_numerator - denominator
@@ -827,7 +837,7 @@ class HSMM:
         bic = []
         keys = list(X.keys())
         models = {
-            f"model_{i}": None for i in range(len(states))
+            f"model_{i}": {} for i in range(len(states))
         }
         for i, n_states in enumerate(states):
             hsmm = GaussianHSMM(n_states=n_states,
@@ -975,7 +985,7 @@ class HSMM:
 
         return RUL, mean_RUL, UB_RUL, LB_RUL
 
-    def prognostics(self, data, max_samples=None, plot_rul=True, get_metrics=True, equation=1):
+    def prognostics(self, data, max_samples=None, plot_rul=True, get_metrics=True, equation=1, return_results=False):
         """
         Performs prognostics for given degradation histories, estimating RUL utilizing the ``RUL`` method and saving the results.
 
@@ -991,10 +1001,24 @@ class HSMM:
             Whether to compute and save evaluation metrics. Default is True.
         equation : int, optional
             Equation type for RUL estimation. Default is 1.
+        return_results : bool, optional
+            Whether to return the results: mean_rul_per_step, pdf_ruls_all, upper_rul_per_step, lower_rul_per_step (default is False).
 
         Returns
         -------
-        None
+        None if return_results is False
+
[…]
