+++ himap/utils.py
@@ -557,7 +557,7 @@ def calculate_cdf(pmf, confidence_level):
     return lower_value, upper_value
 
 
-def create_folders():
+def create_folders(results_parent_path):
     """
     Create a directory structure for storing results.
 
@@ -596,11 +596,14 @@ def create_folders():
             print(f"Folder already exists: {path}")
 
     # Create folders
-    folder_path = os.path.join(os.getcwd(), "results")
+    if results_parent_path is None:
+        folder_path = os.path.join(os.getcwd(), "himap_results")
+    else:
+        folder_path = os.path.join(results_parent_path, "himap_results")
     create_folder(folder_path)
 
     subfolder_names = ["dictionaries", "figures", "models"]
 
     for subfolder_name in subfolder_names:
-        subfolder_path = os.path.join(os.getcwd(), "results", subfolder_name)
+        subfolder_path = os.path.join(folder_path, subfolder_name)
         create_folder(subfolder_path)
