### 2026-04-24 #9491 @pdebuyl
Yes, I was referring to a misunderstanding that the CLI was only meant as a sort of demonstrator. One sentence in the article + one in the README would be fine.

PS: sorry, the conclusion of this review arrived during a job change. This CLI usage clarification is the only thing left before I approve and forward to the EiCs.
