### 2026-04-29 #9491 @mariana-sc
Hi @pdebuyl, we have addressed the minor issues you mentioned and pushed a new release. Please set the version to 1.4.0. Let us know if everything is in order. Thank you!
