### 2026-05-01 #9491 @pdebuyl
Thanks @mariana-sc and others for you interest in JOSS! Thanks @kyleniemeyer for publication.
